Fix to do proper removal of "dead" transaction's Nodes so they don't get left dangling wasting space. More refactors.
